requirements.txt
torch
transformers
pandas
numpy
scikit-learn
tqdm
streamlit
PyYAML
